requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
cffi==1.16.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
setuptools==69.0.3
wheel==0.42.0
yarl==1.9.4
youtube-dl==2021.12.17
